+++ package.json
@@ -14,6 +14,7 @@
   "author": "Devish Baidawar Chhetri",
   "license": "ISC",
   "dependencies": {
+    "ejs": "^3.1.5",
     "eslint": "^6.6.0",
     "express": "^4.17.1",
     "nodemon": "^1.19.4"
